Added filtering by disorder regions using IUPred3

+++ my_modules/ELM_motif_query/modules/motif_disordered_plot.py
@@ -6,13 +6,14 @@ import matplotlib.pyplot as plt
 import subprocess
 import os
 
-def disordered_plot():
-        fig, ax = plt.subplots(figsize=(17,5))
+def disordered_plot(scores):
+    fig, ax = plt.subplots(figsize=(17,5))
     fontsize = 14
 
     # Plot IUPred2 scores returned by API
-    scores = r.json()["iupred2"]
-    ax.plot(np.arange(len(scores)+1)[1:], scores, lw=2)
+    scores = np.asarray(scores, dtype=float)
+    x = np.arange(1, len(scores) + 1)
+    ax.plot(x, scores, lw=2)
 
     # Set x and y axis labels
     ax.set_xlabel("Residue", fontsize=fontsize)
@@ -28,62 +29,48 @@ def disordered_plot():
     ax.grid(True, which="both", color="lightgray", ls="--", lw=1)
     ax.set_axisbelow(True)
 
-    fig.show()
+    fig.show() # use plt.show() if doesnt work
+    return fig, ax
 
-def find_disorder_regions(sequence, iupred_type="long"):
+def run_iupred(sequence, iupred_type="long"):
 
     with open("temp.seq", "w") as f:
         f.write(sequence)
     
-    result = subprocess.run(
-        ["python", "iupred3.py", "temp.seq", iupred_type],
-        capture_output=True,
-        text=True
-    )
+    try:
+      result = subprocess.run(
+          ["python", "iupred3.py", "temp.seq", iupred_type],
+          capture_output=True,
+          text=True)
+      return result.stdout
 
-    os.remove("temp.seq")
-    return result.stout
-
-    # Get amino acid positions of all disordered residues (thresholded as IUPred score > 0.5)
-    disordered = np.arange(len(scores)+1)[1:][np.array(scores) > 0.5]
-
-    # Get amino acid positions of all ordered residues (thresholded as IUPred score < 0.5)
-    ordered = np.arange(len(scores)+1)[1:][np.array(scores) < 0.5]
+    finally:
+      if os.path.exists("temp.seq"):
+        os.remove("temp.seq")
 
 def is_subset(arr1, arr2):
-  """
-  Function to check if an array (arr2) is a subset of anoter array (arr1).
+  return set(arr2).issubset(set(arr1))
 
-  Rerturns True is arr2 is a subset of arr1.
-  """
-  m = len(arr1)
-  n = len(arr2)
-  s = set()
-  for i in range(m):
-      s.add(arr1[i])
+# Get amino acid positions of all disordered residues (thresholded as IUPred score > 0.5)
+    disordered = np.arange(len(scores)+1)[1:][np.array(scores) > 0.5]
 
-  p = len(s)
-  for i in range(n):
-      s.add(arr2[i])
-
-  if (len(s) == p):
-    return True
-  else:
-    return False
+# Get amino acid positions of all ordered residues (thresholded as IUPred score < 0.5)
+    ordered = np.arange(len(scores)+1)[1:][np.array(scores) < 0.5]
 
 # Categorize each motif based on the IUPred scores of its residues
 
 iupred = []
 for index, row in regex.iterrows():
   # Get the start and end amino acid positions of the motif
-  start = row['motif_start_in_query']
-  end = row['motif_end_in_query']
+  start = int(row['motif_start_in_query'])
+  end = int(row['motif_end_in_query'])
+  motif_postions = range(start, end + 1)
 
   # Check if all positions covered by the motif are within ordered or disordered IUPred scores thresholds
-  if is_subset(ordered, list(range(start, end+1))):
+  if is_subset(ordered, list(motif_postions)):
     iupred.append("ordered")
 
-  elif is_subset(disordered, list(range(start, end+1))):
+  elif is_subset(disordered, list(motif_postions)):
     iupred.append("disordered")
 
   else:
